requirements.txt
appdirs==1.4.4
    # via black
appnope==0.1.2
    # via ipython
attrs==20.3.0
    # via
    #   -r requirements.in
    #   pytest
backcall==0.2.0
    # via ipython
black==20.8b1
    # via -r requirements.in
click==7.1.2
    # via
    #   black
    #   pip-tools
coverage==5.4
    # via pytest-cov
decorator==4.4.2
    # via ipython
flake8==3.8.4
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
ipdb==0.13.4
    # via -r requirements.in
ipython-genutils==0.2.0
    # via traitlets
ipython==7.20.0
    # via ipdb
isort==5.7.0
    # via -r requirements.in
jedi==0.18.0
    # via ipython
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
mypy==0.800
    # via -r requirements.in
packaging==20.9
    # via pytest
parso==0.8.1
    # via jedi
pathspec==0.8.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==5.5.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.16
    # via ipython
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pycodestyle==2.6.0
    # via flake8
pyflakes==2.2.0
    # via flake8
pygments==2.8.0
    # via ipython
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements.in
pytest-mock==3.5.1
    # via -r requirements.in
pytest==6.2.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-mock
regex==2020.11.13
    # via black
toml==0.10.2
    # via
    #   black
    #   pytest
traitlets==5.0.5
    # via ipython
typed-ast==1.4.2
    # via
    #   black
    #   mypy
typing-extensions==3.7.4.3
    # via
    #   black
    #   mypy
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
